requirements.txt
moralis==0.1.49
openai==1.61.0
python-dotenv==1.0.1
gradio==5.15.0
scikit-learn==1.6.1
sentence-transformers==3.4.1